# specs/openapi.yaml
openapi: 3.1.0
info:
  title: BenefitServicing Workbench — Command API
  version: 1.0.0
  description: >
    Authoritative machine-readable contract for the BenefitServicing Workbench backend.
    The API exposes **business commands**, not generic document mutations (see specs/02 P1,
    specs/11). Reads are Firestore client-SDK subscriptions to read models (specs/05) and are
    NOT part of this API, except the few server-composed GETs and the health probes.


    Conventions (specs/11, specs/README):
      - All money is integer US cents (int64).
      - Every mutating command requires an `Idempotency-Key` header (specs/08).
      - Optional `If-Match` carries the caller's expected `revision` for stale-UI protection.
      - A *failed payment* is a completed command → HTTP 200 with `contribution.status = FAILED`,
        NOT an HTTP error. HTTP errors mean the command did not complete.
  license:
    name: Proprietary (demo)
servers:
  - url: https://api.demo.example/api/v1
    description: demo
  - url: http://localhost:8000/api/v1
    description: local
security:
  - bearerAuth: []
...
paths:
  /benefit-agreements/{agreementId}/activate:
    post:
      tags: [benefit-agreements]
      summary: Activate a pending benefit agreement (async schedule generation)
      description: >
        Role SERVICING_MANAGER+. Transactionally moves the agreement PENDING→ACTIVATING,
        solves installment amounts so Σ == totalCommitment (specs/07 §7.3), and enqueues the
        schedule-generation task (specs/10 §10.1). Returns 202 on first acceptance; a replay of
        a completed activation returns 200 with the ACTIVE agreement.
      operationId: activateBenefitAgreement
      parameters:
        - $ref: '#/components/parameters/AgreementId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivateRequest' }
      responses:
        '202': { $ref: '#/components/responses/OperationAccepted' }
        '200':
          description: Idempotent replay of an already-completed activation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BenefitAgreement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /benefit-agreements/{agreementId}/suspend:
    post:
      tags: [benefit-agreements]
      summary: Suspend an active benefit agreement
      description: Role SERVICING_MANAGER+. ACTIVE→SUSPENDED, sets acceptingPayments=false (specs/10 §10.2).
      operationId: suspendBenefitAgreement
      parameters:
        - $ref: '#/components/parameters/AgreementId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200': { $ref: '#/components/responses/BenefitAgreementResult' }
        '202': { $ref: '#/components/responses/OperationInProgress' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /benefit-agreements/{agreementId}/resume:
    post:
      tags: [benefit-agreements]
      summary: Resume a suspended benefit agreement
      description: Role SERVICING_MANAGER+. SUSPENDED→ACTIVE, sets acceptingPayments=true.
      operationId: resumeBenefitAgreement
      parameters:
        - $ref: '#/components/parameters/AgreementId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200': { $ref: '#/components/responses/BenefitAgreementResult' }
        '202': { $ref: '#/components/responses/OperationInProgress' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /benefit-agreements/{agreementId}/terminate:
    post:
      tags: [benefit-agreements]
      summary: Terminate a benefit agreement (cancels future contributions async)
      description: >
        Role SERVICING_MANAGER+. ACTIVE/SUSPENDED→TERMINATED (terminal), acceptingPayments=false,
        enqueues cancel-future-contributions (specs/10 §10.3–10.4). In-flight PROCESSING
        contributions settle then cancel; they are never force-cancelled.
      operationId: terminateBenefitAgreement
      parameters:
        - $ref: '#/components/parameters/AgreementId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200': { $ref: '#/components/responses/BenefitAgreementResult' }
        '202': { $ref: '#/components/responses/OperationInProgress' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /borrowers/{borrowerId}/employment-status:
    post:
      tags: [employment]
      summary: Change a borrower's employment status (triggers benefit cascade)
      description: >
        Role SERVICING_MANAGER+. Transactionally updates employment + cascades the benefit
        (LEAVE→suspend, TERMINATED→terminate+cancel-future, ACTIVE→resume) — specs/10 §10.4.
        Requires UI confirmation before submit.
      operationId: changeEmploymentStatus
      parameters:
        - $ref: '#/components/parameters/BorrowerId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmploymentStatusChangeRequest' }
      responses:
        '200':
          description: Employment updated; benefit cascade applied. Future-contribution cancellation may still be running.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmploymentChangeResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: Firebase Authentication ID token, verified by Django via the Admin SDK (specs/12).
  parameters:
    AgreementId:
      name: agreementId
      in: path
      required: true
      schema: { type: string }
      example: ben_jordan_001
    BorrowerId:
      name: borrowerId
      in: path
      required: true
      schema: { type: string }
      example: bor_jordan
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key making the command safe to retry exactly once (specs/08).
      schema: { type: string, minLength: 8, maxLength: 200 }
      example: process-ben_jordan_001__004-a1b2c3
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: The caller's expected `revision` of the target entity; mismatch → 409 STALE_WRITE.
      schema: { type: string }
      example: "12"
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Trace id; generated if absent and echoed into events/logs.
      schema: { type: string }
  responses:
    OperationAccepted:
      description: >
        Operation accepted; asynchronous work in progress. Poll by re-POSTing the identical
        request with the SAME Idempotency-Key (202 while running, final result when done), or
        observe the entity document via Firestore subscription (recommended for the UI).
        Never re-issue under a new key.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Suggested seconds before polling.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OperationStatus' }
    OperationInProgress:
      description: >
        An operation for this key/entity is still running. Poll by re-POSTing the identical
        request with the SAME Idempotency-Key, or observe the entity via subscription.
        Any mutating endpoint may return this. Never re-issue under a new key.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/OperationStatus' }
    BenefitAgreementResult:
      description: The updated benefit agreement.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/BenefitAgreement' }
...
    BadRequest:
      description: Validation error (e.g. missing Idempotency-Key, empty note, bad enum).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid Firebase ID token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but the role is not permitted this command.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: >
        Typed conflict: INVALID_TRANSITION, INVARIANT_VIOLATION, IDEMPOTENCY_KEY_REUSED,
        STALE_WRITE, BENEFIT_NOT_ACCEPTING_PAYMENTS.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unprocessable:
      description: Well-formed but business-invalid (e.g. activate when employment not ACTIVE).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    BenefitStatus:
      type: string
      enum: [DRAFT, PENDING, ACTIVATING, ACTIVE, SUSPENDED, COMPLETED, TERMINATED]
    EmploymentStatus:
      type: string
      enum: [PENDING, ACTIVE, LEAVE, TERMINATED]
...
    ErrorCode:
      type: string
      enum:
        - IDEMPOTENCY_KEY_REQUIRED
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        - INVALID_TRANSITION
        - INVARIANT_VIOLATION
        - IDEMPOTENCY_KEY_REUSED
        - STALE_WRITE
        - BENEFIT_NOT_ACCEPTING_PAYMENTS
        - UNPROCESSABLE

    # ---------- shared value objects ----------
    Cents:
      type: integer
      format: int64
      description: US cents (integer). No floating point.
      example: 83333
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            correlationId: { type: string }
    OperationStatus:
      type: object
      required: [state, operation]
      properties:
        state:
          type: string
          enum: [ACCEPTED, IN_PROGRESS, COMPLETED, FAILED]
        operation: { type: string, example: ACTIVATE_BENEFIT }
        entityType: { type: string }
        entityId: { type: string }
        retryAfterSeconds: { type: integer }
        correlationId: { type: string }
...
    BenefitAgreement:
      type: object
      properties:
        id: { type: string, example: ben_jordan_001 }
        borrowerId: { type: string }
        borrowerName: { type: string }
        employerId: { type: string }
        employerName: { type: string }
        loanId: { type: string }
        currency: { type: string, const: USD }
        totalCommitmentCents: { $ref: '#/components/schemas/Cents' }
        baseMonthlyContributionCents: { $ref: '#/components/schemas/Cents' }
        termMonths: { type: integer }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        amountPaidCents: { $ref: '#/components/schemas/Cents' }
        remainingCommitmentCents: { $ref: '#/components/schemas/Cents' }
        status: { $ref: '#/components/schemas/BenefitStatus' }
        acceptingPayments: { type: boolean }
        suspendedReason:
          oneOf:
            - type: string
              enum: [LEAVE, MANUAL]
            - type: 'null'
        scheduleGenerated: { type: boolean }
        plannedInstallmentCount: { type: integer }
        installmentsGenerated: { type: integer }
        revision: { type: integer }
...
    EmploymentChangeResult:
      type: object
      properties:
        borrowerId: { type: string }
        employmentStatus: { $ref: '#/components/schemas/EmploymentStatus' }
        employmentEndDate: { type: [string, 'null'], format: date-time }
        benefit:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/BenefitAgreement'
        futureCancellation:
          type: object
          properties:
            state: { type: string, enum: [NOT_APPLICABLE, IN_PROGRESS, COMPLETED] }
        correlationId: { type: string }

    # ---------- request bodies ----------
    ActivateRequest:
      type: object
      properties:
        startDate:
          type: string
          format: date
          description: Optional override; defaults to the agreement's configured start date.
        expectedRevision: { type: integer }
    ReasonRequest:
      type: object
      properties:
        reason: { type: string }
        effectiveDate: { type: string, format: date }
        expectedRevision: { type: integer }
    EmploymentStatusChangeRequest:
      type: object
      required: [status, effectiveDate]
      properties:
        status:
          type: string
          enum: [ACTIVE, LEAVE, TERMINATED]   # PENDING is never a legal target (specs/06 §6.5)
        effectiveDate: { type: string, format: date, example: "2026-10-14" }
        reason: { type: string, example: "Employment ended" }
        expectedRevision: { type: integer }
...